continue to fix the bug that current Vomnibar tab may keep active
a race condition was seen
even when 53b96effd05ce987cfbd179ac7cae59fc07606c6 was been applied.
(maybe, I saw it on the first tab of the only Chrome window,
    so the truth has always been a mystery)

According to tests, if FrameMask.Focus focuses at once,
it depends on the order between calling @doEnter and posting "refocusCurrent"
  whether the new tab loses tab focus.
So delay @doEnter for a while seems reasonable.