iframe omnibar: completely fix a bug of width on C53
The commit 0e92706643b7bb2b7d6e4d2682eaae698468da6d is not correct,
because the issue it wanted to fix is only on C53,
so that commit breaks width value on Chrome <= 52.

This function computes the pixel ratio of real screen after a while,
so that real width can be defined well.