chrome version: note a fix on C68
1.
The `LATEST_TESTED` tag on `MinDevicePixelRatioImplyZoomOfDocEl` is a typo,
because once it was ver 66 and then modified into 68 without checks.

A re-checking on Chrome 68.0.3440.75 x64 on Win 10 confirms
that the issue had been fixed.

2.
Update usage of VDom.specialZoom_.

Now tend to think a same zoom as devPixRatio is just a mirror
    (but not configured by a CSS zoom property).