extend click: a work around for elements moving across frames
On QQ Mail's webpage of the desktop version,
elements created by the the top window are adopted into #mainFrame,
but its JS wrapper keeps to be "created by the top window",
which means its aEL is `window.top.addEventListener`.
Therefore, the old logic would throw these elements.

This commit trys fixing this issue.

On Firefox 66/67 (stable) this issue can not be reproduced.